Scaffolding for auth (#54)
- Add auth for service collection. Currently commented out since we
don't have any endpoints that need it yet.
- Note that this setup is specific to Auth0 or similar service as an
authority. The JWT is validated by the authority, not by the
application. If the JWT needed to be verified by us instead, we would
need a slightly different setup.
- Since this adds new named settings, added logic for typed settings.
Can be expanded further as needed, and individual Modules can establish
their own options.
- Renamed `BiblioNexus` connection string to `BiblioNexusDb` for added
clarity (made the change in this PR since I was already adding the typed
config). I've added the new value in Azure, so after deployments can
delete the old one.
- Added an example appsettings file and instructions in the readme.