Add Swagger/OpenAPI support (#57)
Add a swagger service. Can be viewed from `/swagger`. More fine-tuned
options can be set per endpoint if desired, but without setting any
additional options it does a good job.

I also added logic for adding an auth token to example requests for
future endpoints that will need it.

It's also worth noting that the size of some of our responses seems to
overwhelm the javascript that's running to display responses, and it can
freeze up the browser.