add new API endpoints for file manager/consume page (#63)
This rethinks the way the requests are handled in an attempt to simplify usage and optimize response
sizes. There are 7 new endpoints to replace existing ones:

- /bibles/language/LANGUAGE_ID returns basic details about the Bible versions for a language
- /bibles/BIBLE_ID/book/BOOK_ID returns info about a given book of a given Bible version,
including URLs for downloading
- /passages/language/LANGUAGE_ID/resource/TYPE returns basic details for all passages tied to the
given resource type with content in the given language
- /passages/PASSAGE_ID/language/LANGUAGE_ID returns content ids for all resources tied to the
given passage whether through verse or overlapping passage, in the given language. Also includes
supporting resources for the given passage.
- /resources/language/LANGUAGE_ID/book/BOOK_ID?resourceType=CBBTER&resourceType=TyndaleBibleDictionary
returns all the content ids and basic info for a given book and language, filtered to the
resources you specified. Groups by chapter.
- /resources/content/CONTENT_ID returns the JSON content for text types or redirects to the CDN
for other types
- /resources/metadata/CONTENT_ID returns the display name and any other relevant metadata for
content