Resources Summary Endpoint (#72)
- Created a separate folder for ResourcesSummary since the Module was
getting too big. The endpoints in the ResourceModule should be pulled
out in a similar way, but I'm not sure the best way to do that yet. I
need to add similar endpoints for the admin resource list, so I think
that will drive how they get reorganized.
- ResourcesSummaryEndpoints is big. It's using raw SQL queries rather
than trying to mess with LINQ here. Entity Framework is not currently
supporting types not registered in a DbSet without some messing around
(it used to support this, very frustrating), so I went ahead and used
the old ADO.NET stuff for now. However, Entity Framework 8 supposedly
will support it again, which comes out next month, so I left the
original code in here commented out. Once .NET 8 is released and we
upgrade, we can remove the ADO.NET and uncomment the EF stuff.
- Originally I wanted to do more of the data manipulation on the
front-end, and I wasn't sending so much down, but that was turning out
to be very difficult (maybe it's my own lack of javascript experience,
but it seemed like the code to get it done was a lot more complex than
doing it in C#). So a lot of the data manipulation is happening on the
server.
- A couple of the methods look fairly similar, and they probably could
be combined to some extent, but they're using separate groups and when I
started to remove some duplication I was realizing I'd probably create
more code than I'd save. Also, there's the argument that they have
separate reasons to change and thus don't violate DRY.
- I've also gotten the impression that this data is going to change
sooner than later, so I don't think there's a good reason to spend any
more time on it. Always happy to hear suggestions to simplify things
though.
- I added OutputCache. It can have defaults set, but for now just
applying it to this new endpoint. It's an in-memory cache, so not
something we should abuse too much. I think it can be hooked up to
something like a redis cache if we wanted to utilize it more down the
road. But for now it's just there for any calls that we don't think need
to hit the database every time.
- I put what's essentially wild-card CORS logic in as well. The API
didn't work on localhost otherwise (surprised no one's been stopped by
this before). I believe this will override the existing logic on the
Azure service (but again, it's no actual change at the moment).
- Sheesh