don't restrict supporting resource lookups parent checks (#74)
Previously on the resources endpoint, when fetching associated
resources, we were filtering the parent resources down to whatever
resourceTypes were passed. In the case of UbsImages (linked to CBBTER
resources), this meant you couldn't just pass UbsImages without also
passing CBBTER or you would get empty results.

By removing the filtering of parent resources, we ensure that regardless
of what type the associated resources are linked to, they'll be
returned.