rename ResourceType to ParentResource (tables and endpoints)
Since the concept of "resource type" is reserved for things like Dictionary vs Study Notes, this
commit updates the DB schema and endpoints to reflect that. What we've been calling "resource type"
up to this point will be called "parent resource", which is the concept of Biblica Study Notes vs
UBS Images.