rename ResourceType to ParentResource (tables and endpoints) (#91)
Since the concept of "resource type" is reserved for things like
Dictionary vs Study Notes, this commit updates the DB schema and
endpoints to reflect that. What we've been calling "resource type" up to
this point will be called "parent resource", which is the concept of
Biblica Study Notes vs UBS Images.

For future reference, the strategy to perform the table renames using
EF's default migration generator was to put `[Table("ParentResources")]`
in front of the `ResourceTypeEntity` class, then generate the
migrations. Then similarly to perform the column rename, put
`[Column("ParentResourceId")]` in front of `ResourceEntity`'s `TypeId`
field and generate the migrations again. This resulted in two separate
migrations to perform the full rename, and at that point I could update
the class/field names to the final result and get rid of the `Table` and
`Column` attributes.