update admin resource endpoints to operate on resource contents instead of resources
This contains breaking changes that need to be deployed alongside changes to the content manager
frontend.

Previously the content manager operated on the resource level, both for listing as well as viewing
and updating.

The new requirements for the content manager make the assumption that more things are happening on
the resource content level (meaning viewing and updating should be per language).

This updates the existing endpoints to fit that model and also adds some additional fields the
frontend will need for the new requirements.

Note that the list endpoint will still operate on the resource level, but now content ids will be
returned alongside the language ids so the client can intelligently link to the right content if a
specific language is being filtered on.