add separate migrations project that has 10 minute db timeout (#106)
This is an idea from a conversation we had a few weeks ago when we ran
into issues with deploying due to the migration not getting application
config. We discussed that the long term best solution was to have a
separate migration project that's used for generating/running the
migrations, so that non-database config values aren't needed at runtime
for the migrations.

A recent migration that copies lots of data from one table to another is
currently timing out and by doing this separate project we can kill two
birds with one stone. The separate migration project is configured with
a 10 minute DB timeout while still keeping the default DB timeout for
the regular API project.j