only return other languages on resource page that match the current content's media type (#109)
When on a resource that has audio and text, this
`OtherLanguageContentIds` was including every media type for the other
languages. This doesn't make sense because the user will want to see the
same media type that they're currently viewing, just in a different
language. This small update to the `Where` clause fixes that.