cleanup Swagger docs and move admin endpoints behind /admin
Our Swagger docs were getting a little messy with some of the defaults that Swagger has. The three main issues were:

Module and Endpoints suffixes were being added to the end of the tags, which means the groupings in Swagger were named oddly.

Almost all of the schemas were suffixed with Response which seemed redundant.

The enums were all represented as ints even though the API would accept the string or int value and the JSON responses would return the string value.

Some things should be renamed and different configs added to Swagger to fix all of this. In addition, some new handling should be added for the BookCode enum specifically. So that it's easier to use as a route param and the Swagger docs properly show the 3-character book code.

Finally the admin endpoints should be moved behind /admin and all /admin routes should be removed from Swagger since they shouldn't be publicly exposed.