cleanup Swagger docs and move admin endpoints behind /admin (#110)
Our Swagger docs were getting a little messy with some of the defaults
that Swagger has. The main issue was:

1. Module and Endpoints suffixes were being added to the end of the
tags, which means the groupings in Swagger were named oddly.
2. The enums were all represented as ints even though the API would
accept the string or int value and the JSON responses would return the
string value.

Some things should be renamed and different config added to Swagger to
fix all of this.

Finally the admin endpoints should be moved behind /admin.