return permissions from /users/self endpoint for frontend SSR logic (#123)
There's a need on the frontend for knowing if a user has permission to
fetch a given resource. If we were just doing a SPA this would be
straightforward because we have access to an Auth0 JS instance. However,
since these fetches are also done during SSR, it would be nice to have
the insight as the page is loading and be able to conditionally make
fetches.

The big use case right now is for the `/admin/users` endpoint, since
that requires `read:values` authorization.