allow editing content that is not published
Currently the published version of a resource is being edited. We want to make sure we're only
editing the draft version. Also this adds permission checks to make sure the user is allowed to edit
content (i.e. they're assigned).

Also, update the `read:values` and `write:values` permissions to be more relevant to our app.