Create common azure credential service (#161)
Per these issues:

https://github.com/Azure/azure-sdk-for-net/issues/29471
https://github.com/Azure/azure-sdk-for-net/issues/26584

connecting via the DefaultAzureCredential is very slow because it tries
multiple options before getting to AzureCli (which is what we're all
using). Specifically, it tries ManagedIdentity first along with retry
logic, and this creates a huge lag when trying to first start the
application locally. This is, however, what our Azure services use. So
if we are in local environment, use CLI cred, otherwise use Managed.