move .editorconfig to root and enforce its settings (#160)
This is the result of updating the .editorconfig, running JetBrains code
formatter against the code, then running `dotnet format`. Though the
JetBrains code formatter is optional, the dotnet built-in formatter
settings will be enforced.

This will catch things that were previously in the codebase like:
```
public void MyFunc() {
  ...
}
```

It will also catch and fix unused `using` statements and a lot more.