Fix resources/content not returning in prod (#258)
In production only this endpoint will timeout if the verse range is too
large. I copied the query that EF was generating and put it in SQL
directly and it would run in prod. Never any issues against dev. So it
is only an issue when running from Entity Framework against prod. I have
no idea why.

I was able to narrow down that removing just one of the `||`s in the
passage query it would work. So there seems to be some weird edge case
bug that we're hitting with the multiple ORs in the Any query. Pulling
all of them out to separate pieces has the SQL do a separate exists on
each of them, and it then runs fine against prod. Still no idea why it
worked against dev db though.


![image](https://github.com/BiblioNexusStudio/aquifer-server/assets/138067194/90338ae7-23d4-4658-a5a7-b82e77e001d1)

@MrGrinst Tagging you for visibility