Support for bulk send to review on manager dashboard (#286)
Some things to note...

- I added the new workflow statuses that are coming up soon, so we
wouldn't need to go back and add them here later.
- I modeled `send-for-review` after `assign-editor` where you can pass
either the single item through the path or send the list of content ids.
- In the content-manager-web I removed the reference to the `/admin`
endpoint, so those can be removed after next prod deploy.
- I needed a way to clear out inline comments on the server, so that's
where all the nonsense with the TiptapModel comes into play. I was
unsure about making it the default, since sometimes we might use the
serialize/deserialize as a means to compare. So that's when things got
weird and I took an approach with generics, but because of the circular
reference the generics became sort of circular too. At first glance it
looks really weird, but I think it's actually a pretty clean way to
accomplish this without creating a whole new set of objects, and it can
be expanded as needed in the future.