Add Enabled flag to Users table, and don't send down or accept disabled users.
Not that for an authenticated endpoint a user could still get through with their JWT from Auth0 if the endpoint didn't make any checks against the Users table. Auth0 has a concept of blocking though. So we should disable users through the API to handle disabling in our database as well as blocking on Auth0 side.