Back-end logic for using resource SortOrder (#305)
Mostly just sending it down since the front-end handles sorting itself
on the dashboard. Project page and Resources needs the sorting on this
side.

Also fixed a separate bug where if a manager aquiferized/translated a
resource that was assigned to them as part of a project, it wouldn't
create a new snapshot like it would if they assigned it to someone else.