Fix issues with snapshot history (#311)
Pull Snapshot history saving out of assignment change saving to be more
explicit and make it harder to introduce bugs.

The problem was that the assigned user or status on the draft version
could change before snapshot was saved, and it would be hidden in saving
the assignment status change, which didn't have any dependency on the
draft version being updated. So I also explicitly require the old user
id and status to be passed to the snapshot history so there's no
confusion, even though it does create some reduancy.