fix project item list for newly created projects (#327)
The problem was that we were relying on the
ResourceContentVersionSnapshots table always having a snapshot for a
resource, but this is a wrong assumption for newly created projects. By
making the snapshot a left join and then falling back to the
ResourceContentVersion for word count if there's no snapshot, we can
ensure all items get displayed.