remove old reports and improve caching (#404)
Old report endpoints are not in use in prod now and can be removed. Also
we had places that were authenticated endpoints using the
`ServeCacheInSeconds` which does not actually work since any requests
with an `Authorization` bypass the cache by default. The function has
been renamed to make that clearer and a couple reports were updated to
use the client-side `ResponseCache` instead.