Fix Projects load time (#433)
This fixes the Projects page load times because it won't pull many
thousands of resources for the count. However, it slows down the other
two pages generally because they have far fewer resources to pull, so
counting in memory was then faster.

I think the broader solution is that the progress bar should lazy load
on the page, because that part of the query is what's slowing everything
down.