Bib 710 modify view edit page for community reviewer users (#455)
These biggest change here was that I decided to add a separate endpoint
for the community users to send their work to be reviewed for
publishers. Due to the constraints of the user, it made more sense to
create the new endpoint rather than modifying the existing
`send-for-publisher` endpoint. Also, it was a good exercise in
understanding the codebase, EF and FastEndpoints.