Hotfix Commit: Add retry logic to open ai calls (#481)
I couldn't reproduce the 500 not showing up to the front-end. It may be
caused by APIM. I hope this just resolves it anyway, because in my own
local apps that do translation it seems like OpenAI is prone to random
failures, and I've implemented similar retry logic to get around it.
Usually the failures are just a one-off and next try is good.