Translate project and resource content via queues and durable functions. (#515)
This PR accomplishes multiple goals, all of which were required in order
to move translations to an async subscriber:
1. Move Tiptap conversion to C#.
[aquifer-tiptap](https://github.com/biblioNexusStudio/aquifer-tiptap/)
allows creation of a `TiptapUtilities.js` file which could be
copy/pasted and run via Microsoft ClearScript's V8 Engine support (as is
currently done in AI.Worker). Instead of copy/pasting, though, this PR
instead pulls in that source code as a node package and builds the
`TiptapUtilities.js` file on every build.
2. Handle multi-step FIA Tiptap content conversion (to/from JSON/HTML).
3. Move the OpenAI translation and prompts fully to the back-end and
modify it to pull down the entire chat instead of streaming, now using
the C# client in the [OpenAI](https://github.com/openai/openai-dotnet)
package (which comes with some nice built-in error retry handling).
Also, translate paragraphs in parallel to improve translation speeds.
4. Handle multi-step FIA Tiptap content translation.
5. Create a C# utility for getting the word count of Tiptap HTML content
now that `TiptapUtilities.js` no longer supports this functionality.
6. Minify HTML before translating it so that we are less likely to hit
translation character limits. This logic already exists in AI.Worker but
it has been updated to be in a utility method with tests.
7. Use Azure durable functions for fan out -> fan in behavior in order
to translate all resources for a project and then perform final project
related updates.
8. Update the translation flow to create an initial (original language)
snapshot before translating and a second translated content snapshot
after translating.
9. Update the translation flow to do appropriate resource content
version status and assignment changes based upon if the translation
began from project start, a publisher using the translation dropdown, or
a community reviewer clicking the translate button.

Breaking change:
* The `ResourceContentVersionMachineContentTranslation` `UserId` column
is now nullable because a back-end process is creating the translations.
We will probably remove this column altogether in a future change. Note
that the InternalApi now can return `int?` for the `UserId` as well
which is a breaking change but which shouldn't have an impact outside of
updating translation ratings.

Example translations:
* One resource by Publisher via translation dropdown:
https://qa.admin.aquifer.bible/resources/366831
* One resource by Community Reviewer via Translate button:
https://qa.admin.aquifer.bible/resources/366301
* Entire project: https://qa.admin.aquifer.bible/projects/1203

Related PRs to use the new `TiptapConverter` from this PR:
* https://github.com/BiblioNexusStudio/ai-worker/pull/1
* https://github.com/BiblioNexusStudio/content-loader/pull/33