Make Subscribers, Publishers, and Managers consistent. (#529)
This PR is almost all renames using automated refactors:
* Name all services that publish messages to a queue
`*MessagePublisher`.
* Use similar method names for publishing: `Publish*Message`.
* Ensure all message models have consistent naming: `Verb*Message`.
* Name all classes that listen to a queue `*MessageSubscriber`.
* Name all timer trigger function classes `*Manager`.
* Move things into appropriate folders.

One thing that did change: The `EmailService` and
`EmailMessagePublisher` no longer share a model because they are no
longer abstracted to be essentially the same thing. I tested this
locally to make sure it still works.

This will change Azure function names because we name them after the
class they are in. I believe that will be fine.