Refactor notification emails to use unique personalizations and one email per To address. (#532)
Changes:
* Notification emails are now sent using the To field instead of the Bcc
field.
* Notification emails now use a reply-to address of
`no-reply@aquifer.bible` (instead of putting that in the To field).
* Notification emails now also support bulk send with unique
personalization content for templated emails.

Additional details:

Initially I was unable to prevent the sending of a single email with
multiple recipients in the To field. I was able to address this by
sending multiple personalizations in the SendGrid email, one per To,
instead of only one personalization with all of the Tos. This also
allows for unique personalization dynamic template data for each To
email address (the third bullet point above). To confirm that this new
personalization worked correctly I updated all notifications emails to
begin with:
```
Recipient Name,

...
```

A byproduct of this change is that now we send up dynamic template data
multiple times, once for each To email address, even though most of it
is redundant. Because of this I added batching to the email send at 1000
emails per SendGrid API call (their limit is 1000) but we could further
reduce this if the JSON request body size to SendGrid ends up being very
large.

Furthermore, testing with multiple email addresses using `+` in the
email address with same base email will not result in all emails being
delivered by email providers even though SendGrid is successfully
sending them. One way around this is to send multiple batches, one for
each unique `+` email address for each base email address, so that email
providers won't consider them the same email due to different
`Message-ID`s. I implemented a solution for this because we commonly use
plus-aliased email addresses in testing.