Bib 809 machine translation scoring (#620)
This PR addresses the need to evaluate the effectiveness of machine
translations and human edits by determining the similarity between two
versions of a resource. To achieve this, we will utilize the Levenshtein
Distance algorithm, a measure of the minimum number of single-character
edits (insertions, deletions, or substitutions) required to change one
string into another.

The PR includes a Levenshtein Distance implementation and the necessary
utility methods to effectively score long resources. A new queue,
generate-resource-content-similarity-score, has been created, along with
a corresponding message, publisher, and subscriber. The message will
contain information about the resource content versions being compared,
including the type of comparison being performed, as designated by one
of the values in the ResourceContentVersionSimilarityComparisonTypes
enum. This allows the subscriber to run the appropriate logic for the
resource content types being compared.