Allow for graceful succeed when no Machine Translation exists (#632)
There should not be an error thrown when a machine translation does not
exist for a Resource Content Version or Snapshot when an attempt was
made to score similarity. There are many resources that are in a working
state that are in this situation. Instead, we should log a warning and
allow the graceful exit of the scoring job.