Add notifications routes to Internal API. (#644)
New routes:
* `GET /notifications`: Gets all notifications for the current user for
the past 30 days with optional paging and filtering parameters.
* `PATCH /notifications/{kind}/{id}`: Mark a notification as
read/unread.

Note that because there is no true notification entity the notification
kind and notification entity ID are used to identify a notification
(e.g. "Comment" and a comment ID of 123). We could introduce a new
notification ID later if desired which would only require modifying the
PATCH route to use that new ID instead of the above combination.

A new DB table, `Notifications`, was added, which only holds
notifications a user has explicitly interacted with by marking them as
read or unread.