Allow translated items to be selected for new projects. (#652)
Improvements:
* Update `GET /resources/untranslated` and `GET
/resources/unaquiferized` to use the `ResourceContentSearchService`.
This required supporting an `IsTranslated` filter and an
`IsNewestResourceContentVersion` filter.
* Update the `ResourceContentSearchService` to support multiple verse ID
ranges (required for the above).
* Update verse ID range calculation for multiple chapters to condense
consecutive distinct chapters into a range. This improves DB query
performance, especially because the front-end will pass "all" chapters
for a book of the Bible as "1, 2, ..., N" instead of "1-N" and each
chapter was turned into its own verse ID range.

Additions:
* Add `GET /resources/translated` to fetch only translated resource
contents that aren't already in a project. I considered merging this new
route into one route that supports untranslated, unaquiferized, and
translated but now that they just call the search service this didn't
duplicate as much code and the responses are slightly different so I
opted for the new endpoint.
* Support `isAlreadyTranslated` as a flag when creating projects. This
defaults to false and thus is backwards compatible upon deployment.
* Ensure the translation subscriber skips translating these already
translated resource contents but still assigns them when the project is
started.

See https://github.com/BiblioNexusStudio/content-manager-web/pull/573
for the front-end changes.