Use previous WHERE EXISTS with OR strategy for verse ID range filtering. (#665)
This reverts back to SQL similar to what we're running in production for
the verse ID filtering strategy using `WHERE (EXISTS ... OR EXISTS
...)`, though production only supports a single `startVerseId` and
`endVerseId`. This PR also supports multiple verse ID ranges.

Testing on my local and in SSMS this is much faster than the multi-join
strategy I copied from the `/untranslated` endpoint.