Bib 853 versification UI updates (#662)
To allow the frontend to properly format the message for the user as per
the ticket, I'm returning a `BaseMapping` in the response. It will be
`null` if the base bible mapping is the same and include the `bookName
chapter:verse` string if it differs.

This will allow the client to easily mark the verse with the required
`*` and have the base mapping available in a tooltip (if desired) for
clarity.

I made this resource on dev/qa to work as a demo/test resource:
`/resources/381634`