Create endpoints and subscriber to handle audio uploads for resource content. (#694)
Details:
1. Add two new endpoints, one to upload a resource content audio file,
the other to generically get the status of uploads (which could also be
used for future upload types like images or videos).
2. Add a new subscriber to process the uploaded audio files.
3. Use a new CDN container, `aquifer-dev-content`, which allows dev/QA
environment uploads to live on the CDN without overwriting production
CDN data.
4. Add configuration for accessing the CDN blob storage in
`aquiferstorage` to Jobs for publishing to the CDN.
5. Add configuration for accessing the blob storage for the current
environment (`aquiferstoragedev` or `aquiferstorageprod`) to Aquifer.API
and Jobs in order to upload the audio file to temp blob storage.
6. Use ffmpeg in the subscriber to do audio processing, including speech
normalization and silence detection. This copies the strategies already
in use in aquifer-utilities and content-loader.
7. Create a new resource content version which references the uploaded
file as well as a status history for the resource content.