Add route to clear caches to Aquifer.API and Aquifer.PublicAPI. (#714)
I tested and both cache clearing scenarios work as expected on both
endpoints. I also confirmed that the swagger docs do not include the new
endpoint and no client code is generated for it in the Public API. The
endpoints are gated by a new Admin ApiKeyScope which, in addition to
everything granted by `All`, also grants access to admin routes.