Support one-to-many versification mapping. (#722)
* Consider verse parts (data was already in the DB) during versification
mapping, which allows a single verse to map to more than one verse. This
update resulted in a breaking change for the versifications endpoint
which is [addressed in the
CMS](https://github.com/BiblioNexusStudio/content-manager-web/pull/609).
* Add `MinVerseNumber` to `BookChapters` which allows the `ENG` data to
know when there is a verse 0 (Psalm superscript).

Example of where this allows two verses in ENG (verse 0 for superscript
and standard verse 1) to map to a single verse (verse 1 only which
includes superscript) in Russian:

![image](https://github.com/user-attachments/assets/6047aa98-266a-4ed7-ae57-e5f51fc15614)

Example of new multi-mapping in Russian from existing Nehemiah data:

![image](https://github.com/user-attachments/assets/36d0d460-d6e7-476b-85c2-6c86bf8f1e16)