Update comment notifications. (#726)
Changes:
* Comment mentions result in a notification.
* Participation in a comment thread results in notifications for future
comments in that thread.
* Users who were previously assigned the resource content will no longer
get notifications for comments. Only the user actively assigned at the
time the comment was made will get a notification.

Unchanged:
* The user who created the comment will never get a notification for
that comment.
* Only `ResourceContentVersion`s in the Draft status will show
notifications.

Implications:
* If a `ResourceContentVersion` is Published then notifications for any
comments while it was in Draft will disappear.
* If a Comment is edited to remove a Mention then the notification for
that Mention will disappear.
* If a Comment is edited to add a Mention then a notification for that
Mention will appear. However, it will be sorted by the time that the
comment was originally _created_, not the time of the update to include
the mention.