Use latest dotnet feature release. (#731)
We previously were using `9.x`
[here](https://github.com/BiblioNexusStudio/aquifer-server/blob/e13ff656b846fb5045c3e3b46c203c6258655d83/.github/workflows/pre-merge.yml#L20)
before the introduction of `global.json`. Using `latestFeature`
([docs](https://learn.microsoft.com/en-us/dotnet/core/tools/global-json))
is the equivalent of `9.0.x`.

We'll want to keep an eye on .net releases because the GHA runner images
only have one 9.x version installed and we incur a 30-60 second build
time penalty if we download and install a different version. Switching
to `latestFeature` let us use the current version on the GHA runner
image but if there is a minor version release we'll need to manually
update the `global.json` file.